### 2019-10-22 #1827 @whedon
[PRE REVIEW]: BioSimSpace: An interoperable Python framework for biomolecular simulation
**Submitting author:** @lohedges (<a href="http://orcid.org/0000-0002-5624-0500">Lester Hedges</a>)
**Repository:** <a href="https://github.com/michellab/BioSimSpace" target ="_blank">https://github.com/michellab/BioSimSpace</a>
**Version:** 2019.2.0
