### 2019-10-22 #1827 @kthyng
@will-rowe Would you be interested in trying this for your first JOSS submission as editor?
