### 2019-10-22 #1827 @kthyng
@lohedges Is this pointing to the incorrect github repo or branch currently? Can you clarify the difference in the branches and which is correct to use for your JOSS submission?
