### 2019-10-22 #1827 @lohedges
Hi @kthyng. Our paper is currently in the `feature-paper` branch of our repository (within a `paper` directory at the top level). Our plan was to merge this into our development branch, `devel`, when ready for publication, then into `master` for our next release. If you need the paper to be accessible from the main branch then I can merge across into `devel` now. Let me know what is preferable. (The `feature-paper` branch is fully up-to-date with `devel` so has all of the latest features.)

(I had mentioned this in our "paper description" section during the submission, but hadn't thought about the paper being processes by a bot :-) )
