### 2019-10-22 acbebfc
Added license badge. ***NO_CI***
