### 2019-10-22 a33097b
Update SIRE citation. ***NO_CI***
