+++ paper/paper.bib
@@ -59,6 +59,9 @@ year = {2005}
 }
 
 @misc{ref-SIRE,
+    author = "{Christopher Woods}",
     title = {Sire: An advanced, multiscale, molecular simulation framework},
-    howpublished = {\url{http://siremol.org}},
+    year = "2013",
+    howpublished = "\url{http://siremol.org}",
+    note = "[Online; accessed 22-October-2019]"
 }
