### 2019-10-22 70ea73f
Formatting tweaks to improve JOSS PDF rendering. ***NO_CI***
