### 2019-10-22 37538ba
Fix author in Sire citation. ***NO_CI***
