+++ paper/paper.bib
@@ -59,7 +59,7 @@ year = {2005}
 }
 
 @misc{ref-SIRE,
-    author = "{Christopher Woods}",
+    author = "{Woods, Christopher J.}",
     title = {Sire: An advanced, multiscale, molecular simulation framework},
     year = "2013",
     howpublished = "\url{http://siremol.org}",
