### 2019-10-22 #1827 @lohedges
Apologies for all the PDF renderings. Figured out how to compile from a custom branch. Just ironing out some minor formatting issues.

Does anyone know how to force a page break? It would be nice to make sure that the code snippets aren't split across multiple pages. (Obviously this could all change following any edits anyway.)
