+++ paper/paper.md
@@ -38,7 +38,7 @@ In most research communities there is not a single, unified software-framework. 
 
 One of the core features of BioSimSpace is the ability to set up and run molecular dynamics (MD) simulations. There are a large number of packages that can run MD for biomolecules and BioSimSpace supports several of these: [AMBER][@ref-AMBER], [GROMACS][@ref-GROMACS], and [NAMD][@ref-NAMD]. BioSimSpace also comes with a bundled dynamics engine, SOMD, which interfaces with the [OpenMM][@ref-OPENMM] toolkit to provide graphics processing unit (GPU) acceleration. (This means that there is always a fall back in case no other MD engines are installed.)
 
-While, broadly speaking, the different MD engines offer a similar range of features, their interfaces are quite different. This makes it hard to take expertise in one package and immediately apply it to another. At the heart of this problem is the incompatibility between the molecular file formats used by the different packages. While they all contain the same information, i.e. how atoms are laid out in space and how they interact with each other, the structure of the files is very different. In order to provide interoperability betwen packages we will need to be able to read and write many different file formats, and be able to interconvert between them too.
+While, broadly speaking, the different MD engines offer a similar range of features, their interfaces are quite different. This makes it hard to take expertise in one package and immediately apply it to another. At the heart of this problem is the incompatibility between the molecular file formats used by the different packages. While they all contain the same information, i.e. how atoms are laid out in space and how they interact with each other, the structure of the files is very different. In order to provide interoperability betwen packages we need to be able to read and write many different file formats, and be able to interconvert between them too.
 
 # Features
 
