+++ paper/paper.md
@@ -52,7 +52,7 @@ As files are parsed, records in those files are assigned to a set of _properties
 
 ![Files are parsed in parallel with the parser that successfully reads the file determining the file format. Once all files are parsed, a lead parser (solid red arrows) constructs the topology of the molecular system. Records within the file, e.g. representing terms in the molecular potential such as bonds, angles, etc., are converted into file format specific representations, then stored internally as properties of the molecule as general algebraic expressions. Parsers that follow add additional information to an existing system. Here the `AmberRst7` parser adds coordinate and simulation box data to the system (dashed blue arrows). The file format associated with the files is also stored as a property of the system so that nodes can always convert back to the original format on write.](figures/fig1.png)
 
-Another feature of our parsers is guaranteed read/write self-consistency. Any file that can be read can also be written, and vice-versa. In addition, when expected molecular information is missing from a file, unless obvious, we don't attempt to guess what it may have been. In this sense our parsers don't attempt to be _too_ clever, which can lead to unexpected behaviour, particulars when information is modified or supplemented behind the user's back.
+Another feature of our parsers is guaranteed read/write self-consistency. Any file that can be read can also be written, and vice-versa. In addition, when expected molecular information is missing from a file, unless obvious, we don't attempt to guess what it may have been. In this sense our parsers don't attempt to be _too_ clever, which can lead to unexpected behaviour, particular when information is modified or supplemented behind the user's back.
 
 The code below shows how to load a set of AMBER format files from a directory:
 
