### 2019-10-22 778de94
Fix figure number. ***NO_CI***
