+++ paper/paper.md
@@ -144,7 +144,7 @@ node.setOutput("minimised", BSS.IO.saveMolecules("minimised",
 node.validate()
 ```
 
-BioSimSpace nodes are flexible in the way in which they can be used, with the same script working seamlessly from within a Jupyter notebook or on the command-line. Typically, a user would a write a node as a fully documented, interactive Jupyter notebook, then save it as a regular Python script to run from the command-line. (For inclusion here we simply include the Python script representation of the node, which could be re-converted to a notebook using, e.g., p2j.) Any purely interactive elements included in the node, e.g.  visualisations and plots, are simply ignored when the script is run in a non-interactive mode. To facilitate this dual-use the `node.addInput` method generates a custom ipywidgets based graphical user interface for interative use in Jupyter, or a custom argparse parser for handling command-line arguments. Figure 1 shows the example node above running within a Jupyter notebook (top) and from the command-line (bottom).
+BioSimSpace nodes are flexible in the way in which they can be used, with the same script working seamlessly from within a Jupyter notebook or on the command-line. Typically, a user would a write a node as a fully documented, interactive Jupyter notebook, then save it as a regular Python script to run from the command-line. (For inclusion here we simply include the Python script representation of the node, which could be re-converted to a notebook using, e.g., p2j.) Any purely interactive elements included in the node, e.g.  visualisations and plots, are simply ignored when the script is run in a non-interactive mode. To facilitate this dual-use the `node.addInput` method generates a custom ipywidgets based graphical user interface for interative use in Jupyter, or a custom argparse parser for handling command-line arguments. Figure 2 shows the example node above running within a Jupyter notebook (top) and from the command-line (bottom).
 
 ![BioSimSpace nodes can be run within a Jupyter notebook (top) or
 from the command-line (bottom)](figures/fig2.png)
