### 2019-10-22 #1827 @kthyng
> Hi @kthyng. Our paper is currently in the `feature-paper` branch of our repository (within a `paper` directory at the top level). Our plan was to merge this into our development branch, `devel`, when ready for publication, then into `master` for our next release. If you need the paper to be accessible from the main branch then I can merge across into `devel` now. Let me know what is preferable. (The `feature-paper` branch is fully up-to-date with `devel` so has all of the latest features.)
>
> (I had mentioned this in our "paper description" section during the submission, but hadn't thought about the paper being processes by a bot :-) )

Actually I did click the button myself (not a bot) and I saw your note but wasn't careful enough and am not sure where to find that information now that I clicked the button to start the issue :) That's how I knew to ask anyway.

As you found, you can generate the pdf from a branch. From my understanding, the reviewers will review the code in the github repo listed on this review issue, but the paper is located in the `feature-paper` branch of this repo. Is that accurate?
