### 2019-10-22 #1827 @kthyng
@lohedges Can you go through your paper and work on how to cite your references inline?  You can see details here: https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography
I'm looking especially at the first paragraph under Molecular dynamics on the first page.
