### 2019-10-22 #1827 @kthyng
@lohedges Re: forcing page break:
I am not sure but this will work in Markdown generally so I think it should work here (found [here](https://stackoverflow.com/questions/22601053/pagebreak-in-markdown-while-creating-pdf)):

`<div style="page-break-after: always;"></div>`
