+++ paper/paper.md
@@ -36,7 +36,7 @@ In most research communities there is not a single, unified software-framework. 
 
 # Molecular dynamics
 
-One of the core features of BioSimSpace is the ability to set up and run molecular dynamics (MD) simulations. There are a large number of packages that can run MD for biomolecules and BioSimSpace supports several of these: [AMBER][@ref-AMBER], [GROMACS][@ref-GROMACS], and [NAMD][@ref-NAMD]. BioSimSpace also comes with a bundled dynamics engine, SOMD, which interfaces with the [OpenMM][@ref-OPENMM] toolkit to provide graphics processing unit (GPU) acceleration. (This means that there is always a fall back in case no other MD engines are installed.)
+One of the core features of BioSimSpace is the ability to set up and run molecular dynamics (MD) simulations. There are a large number of packages that can run MD for biomolecules and BioSimSpace supports several of these: AMBER [@ref-AMBER], GROMACS [@ref-GROMACS], and NAMD [@ref-NAMD]. BioSimSpace also comes with a bundled dynamics engine, SOMD, which interfaces with the OpenMM [@ref-OPENMM] toolkit to provide graphics processing unit (GPU) acceleration. (This means that there is always a fall back in case no other MD engines are installed.)
 
 While, broadly speaking, the different MD engines offer a similar range of features, their interfaces are quite different. This makes it hard to take expertise in one package and immediately apply it to another. At the heart of this problem is the incompatibility between the molecular file formats used by the different packages. While they all contain the same information, i.e. how atoms are laid out in space and how they interact with each other, the structure of the files is very different. In order to provide interoperability betwen packages we need to be able to read and write many different file formats, and be able to interconvert between them too.
 
@@ -44,7 +44,7 @@ While, broadly speaking, the different MD engines offer a similar range of featu
 
 ## Parsers
 
-At its core, BioSimSpace is built around a powerful set of file parsers which allow reading and writing of a wide range of molecular file formats. File input/output is provided via the `BioSimSpace.IO` package using parsers from the [Sire][@ref-SIRE] molecular simulation framework, on top of which BioSimSpace is built. Unlike many other programs, we take the approach that it is the _contents_ of the file that defines it format, not the _extension_. As such, we attempt to parse a file with all of our parsers in parallel. Any parser for which the contents of the file is incompatible will be rejected early, with the eventual format of the file determined by the parser that completed without error.
+At its core, BioSimSpace is built around a powerful set of file parsers which allow reading and writing of a wide range of molecular file formats. File input/output is provided via the `BioSimSpace.IO` package using parsers from the Sire [@ref-SIRE] molecular simulation framework, on top of which BioSimSpace is built. Unlike many other programs, we take the approach that it is the _contents_ of the file that defines it format, not the _extension_. As such, we attempt to parse a file with all of our parsers in parallel. Any parser for which the contents of the file is incompatible will be rejected early, with the eventual format of the file determined by the parser that completed without error.
 
[…]
