### 2019-10-23 #1827 @lohedges
@kthyng

> Actually I did click the button myself (not a bot) and I saw your note but wasn't careful enough and am not sure where to find that information now that I clicked the button to start the issue :) That's how I knew to ask anyway.

Apologies for confusing you with a bot :-) You are indeed a _bot master_, wheddon caters to _your_ whims, not vice versa.

> From my understanding, the reviewers will review the code in the github repo listed on this review issue, but the paper is located in the feature-paper branch of this repo. Is that accurate?

Yes, the repository is located at the link given at the top of this thread, i.e. [https://github.com/michellab/BioSimSpace](https://github.com/michellab/BioSimSpace). This points to the default branch of our repo, which is `devel`, which is what should be used for review purposes. However, the paper is currently located on the `feature-paper` branch, so reviewers will currently need to use that branch when generating a PDF. (The branch is fully up to date with `devel`, so there is no issue with reviewing the code from either branch, but ultimately the repository link that accompaines the paper should be the one listed above.) As mentioned, the plan is to merge `feature-paper` into `devel` when the paper is ready for publication, then `devel` will be merged into `master` at the time of our next release.

> Can you go through your paper and work on how to cite your references inline? You can see details here: https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography

This is now fixed. I'm not quite sure how that happened, since I followed the guide you linked to when writing the paper. Originally they were hyperlinks to the websites of the different molecular dynamics engines, so I must have forgotten to delete the first set of square brackets.

> I am not sure but this will work in Markdown generally so I think it should work here (found here):

Thanks, I've tested locally and this seems to work fine. I'll leave as is for now since things will likely jump around a bit if edits are needed. I'll make sure to tidy up the layout as much as possible at the proof stage.

Thanks for your help, much appreciated.
