### 2019-10-23 #1827 @lohedges
Thanks, @will-rowe. I'm happy for the review process to start.

Let me know if you, or the reviewers, have any questions about the repository layout that I discussed in reply to @kthyng above. I understand that this is possibly non-standard, but hopefully it doesn't cause issue with the review process.

Cheers!
