### 2019-10-23 1c2cefc
Added JOSS status badge. ***NO_CI***
