### 2019-10-23 #1831 @will-rowe
Hi @amritagos and @richardjgowers - thank you both for agreeing to review!

In case you missed it on the pre-review issue, the author has some instructions as to branches for the review:

*  the repository is located at ... https://github.com/michellab/BioSimSpace. This points to the default branch of our repo, which is devel, which is what should be used for review purposes. However, the paper is currently located on the feature-paper branch, so reviewers will currently need to use that branch when generating a PDF.
