### 2019-10-23 #1831 @will-rowe
Please carry out your review in this issue by updating the checklist above and giving feedback in this issue. The reviewer guidelines are available [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html)
