### 2019-10-24 michellab/BioSimSpace#128 @lohedges
Lester's TODO list
Just a check list of known issues and things TODO before I go on leave:

* [ ] 1) Finish JOSS submission process. Review thread is [here](https://github.com/openjournals/joss-reviews/issues/1831).

* [x] 2) Document versions of external tools that BioSimSpace supports. Probably hard to be comprehensive, but can at least document versions that we've tested against, e.g. GROMACS > 5.0, AmberTools18/AmbertTools19, etc.

* [ ] 3) Updated `BioSimSpace.Process.Amber` to add support for AMBER simulations using `pmemd` and `pmemd.cuda`. Currently we only support `sander`. (`pmemd` and `pmemd.cuda` are searched for when using the `BioSimSpace.MD.run` function.)

* [ ] 4) Handle execution of MD processes using `mpirun`. Probably need to detect the presence of a hostfile. Will likely need to support `mpirun`, `mpiexec`, and `srun`. (Note that `srun` doesn't require a hostfile, so there might not be a single solution.)

* [x] 5) Make error messages more informative for some of the common exceptions that are trigged. It could be best to add a _verbose_ mode that prints the full Sire error if activated.

* [ ] 6) Work out how to capture errors within a node. I think we might need something like `node.addProcess`, which will zip up the working directory of a process if it fails and present this to the user, e.g.

```python
# Generate an MD process to run the protocol.
process = BSS.MD.run(system, protocol)

# Add the process to the node so that error output can be captured.
# Could be tricky if the user overwrites the process within the same node.
node.addProcess(process)
```

* [x] 7) Add documentation explaining how to modify a protocol. This is a little tricky since protocols aren't defined in isolation and can require information from multiple objects, e.g. the `System` and the `Process`. It should be possible to explain the basic concepts though. Perhaps it is easiest for the user to override the `_generate_config` methods of each process to customise the protocols.
