### 2019-10-24 michellab/BioSimSpace#128 @jmichel80
0. Finish the JOSS submission process should stay on top of the to-do list.
1.  Yes important.
2. Yes we should run pmemd.cuda by default on GPU instances for plain MD (if available).
3. Sounds tricky to make it work reliably and test thoroughly in a short amount of time?
4. If you list the exceptions you have in mind we can discuss what a useful error message would look like for each of those instances.
5. Sounds useful but I worry it's harder than it looks and therefore not doable within the time left.
6. That's important, we should aim to describe well the process to implement the currently implemented different protocols (MD or free energy or metadynamics...) so the doc can be used as a roadmap for future protocols. I would do that rather than refactor code to make the job easier for the user.
