### 2019-10-24 michellab/BioSimSpace#128 @lohedges
I've reverted the original formatting in my first post since it was specific markdown so I could check items off as they are completed. I've added your numbering _after_ the square brackets instead.

Thanks for the comments. This is also list of known limitations as well as things to do, just so whoever is maintaining things is aware of them in my absence.

With regards to the JOSS submission: Since the review process is now active, there's not much I can do until the reviews are in. This _should_ be done within two weeks (hopefully sooner) so these were to keep me occupied in the mean time.

For 5), I've now implemented a _verbose_ mode that prints the full exception trace if activated. This can be done using `BSS.setVerbose(True)` if working interactive or in a script, or using `--verbose` when running a node from the command-line. This just prints the full traceback, which is mostly the output of whatever Sire exception was triggered. I'll push this to `devel` later when I have tested it fully. I'll also try to group some common exceptions and post them here. (Most are IO errors). I feel like I've already done a pretty good job of catching errors that occur during merging, or when copying properties from a parameterised molecule back into the molecule that was originally passed in.
