### 2019-10-24 7b5e5b8
Added "verbose" mode to enable access to full error trace. [ref #128]
