+++ python/BioSimSpace/Process/_somd.py
@@ -38,6 +38,7 @@ import warnings as _warnings
 from Sire import Base as _SireBase
 from Sire import IO as _SireIO
 
+from BioSimSpace import _isVerbose
 from BioSimSpace._Exceptions import IncompatibleError as _IncompatibleError
 from BioSimSpace._Exceptions import MissingSoftwareError as _MissingSoftwareError
 from BioSimSpace._SireWrappers import System as _System
@@ -259,15 +260,23 @@ class Somd(_process.Process):
         try:
             rst = _SireIO.AmberRst7(system._sire_object, self._property_map)
             rst.writeToFile(self._rst_file)
-        except:
-            raise IOError("Failed to write system to 'RST7' format.") from None
+        except Exception as e:
+            msg = "Failed to write system to 'RST7' format."
+            if _isVerbose():
+                raise IOError(msg) from e
+            else:
+                raise IOError(msg) from None
 
         # PRM file (topology).
         try:
             prm = _SireIO.AmberPrm(system._sire_object, self._property_map)
             prm.writeToFile(self._top_file)
-        except:
-            raise IOError("Failed to write system to 'PRM7' format.") from None
+        except Exception as e:
+            msg = "Failed to write system to 'PRM7' format."
+            if _isVerbose():
+                raise IOError(msg) from e
+            else:
+                raise IOError(msg) from None
 
         # Generate the SOMD configuration file.
         # Skip if the user has passed a custom config.

[…]
