### 2019-10-28 96c6353
Added "compatibility" section to website. [ref #128]
