+++ doc/source/compatibility.rst
@@ -1,0 +1,27 @@
+.. _ref_compatibility:
+
+=============
+Compatibility
+=============
+
+BioSimSpace has been tested against the following versions of its
+external dependencies:
+
+
+`AMBER <http://ambermd.org>`__
+==============================
+
+* AmberTools19
+* AmberTools18
+
+`GROMACS <http://www.gromacs.org>`__
+====================================
+
+* 2019 series
+* 2018 series
+
+`NAMD <https://www.ks.uiuc.edu/Research/namd>`__
+================================================
+
+* 2.13
+* 2.12
