+++ doc/source/nodes.rst
@@ -196,4 +196,3 @@ Examples
 ========
 
 A set of example nodes are available on our :ref:`notebook server <ref_notebook>`.
-
