+++ doc/source/index.rst
@@ -3,6 +3,18 @@
 BioSimSpace
 ===========
 
+.. image:: https://dev.azure.com/michellab/BioSimSpace/_apis/build/status/michellab.BioSimSpace?branchName=devel
+   :target: https://dev.azure.com/michellab/BioSimSpace/_build
+   :alt: Build Status
+
+.. image:: https://anaconda.org/michellab/biosimspace/badges/downloads.svg
+   :target: https://anaconda.org/michellab/biosimspace
+   :alt: Conda Downloads
+
+.. image:: https://img.shields.io/badge/License-GPL%20v2-blue.svg
+   :target: https://www.gnu.org/licenses/old-licenses/gpl-2.0.en.html
+   :alt: License
+
 `BioSimSpace <https://github.com/michellab/BioSimSpace>`__ is an interoperable
 Python framework for biomolecular simulation. With it you can:
 
