### 2019-10-28 b305b3b
Added JOSS status badge to website.
