+++ doc/source/index.rst
@@ -15,6 +15,10 @@ BioSimSpace
    :target: https://www.gnu.org/licenses/old-licenses/gpl-2.0.en.html
    :alt: License
 
+.. image:: https://joss.theoj.org/papers/4ba84ad443693b5dded90e35bf5f8225/status.svg
+   :target: https://joss.theoj.org/papers/4ba84ad443693b5dded90e35bf5f8225
+   :alt: Paper
+
 `BioSimSpace <https://github.com/michellab/BioSimSpace>`__ is an interoperable
 Python framework for biomolecular simulation. With it you can:
 
