### 2019-10-29 654ca1b
Explicitly mention CCPBioSim in summary/acknowledgments. [ref openjournals/joss-reviews#1831] ***NO_CI***
