### 2019-10-29 b0f56be
Standardise syntax and improve setting of member data. ***NO_CI***
