### 2019-11-09 michellab/BioSimSpace#129 @amritagos
[JOSS Review] Author list
This is part of https://github.com/openjournals/joss-reviews/issues/1831.

 I notice that there are [five direct contributors](https://github.com/michellab/BioSimSpace/graphs/contributors) to the repository and four authors for the paper. [It appears that](https://github.com/michellab/BioSimSpace/commits?author=ptosco) @ptosco added some tests back in August. The question is regarding the author list then. Is there a reason to not include the fifth contributor (can't be contacted, left the project before the paper was begun, etc)?
