### 2019-11-09 michellab/BioSimSpace#129 @lohedges
Hi @amritagos. @ptosco is not a direct member of the core BioSimSpace team (who are the authors of the paper) but works for an external industrial project partner (Cressett). You are correct that he added some tests and bug fixes over the summer. Since BioSimSpace is built on top of [Sire](https://github.com/michellab/Sire) he has also made some important indirect contributions (such as adding windows support).

I'm not sure of the JOSS policy, but should every contribution to the code (regardless of number of commits) result in authorship? I would be happy to have @ptosco as an author if needed. Pinging @jmichel80 and @chryswoods for comments/suggestions.
