### 2019-11-10 michellab/BioSimSpace#129 @amritagos
No JOSS policy just requires that the reviewers are satisfied with the author list. Thank you for the clarifications. Since he is not a direct member and is not actively working with the project your reasoning makes sense.
