### 2019-11-10 michellab/BioSimSpace#130 @richardjgowers
[joss] example notebooks
re: openjournals/joss-reviews#1831

The examples in the paper are pretty long, they could be squashed down a little by getting rid of the redundant comments, ie lines like addAuthor are self documenting.  Other lines could be grouped together with a common comment and less whitespace between
