### 2019-11-10 michellab/BioSimSpace#131 @richardjgowers
[joss] interoperability refs
re: openjournals/joss-reviews#1831

There doesn't seem to be much mention of existing work in interoperability. I'd expect mentions to projects such as parmed, intermol and openff toolkit
