### 2019-11-10 michellab/BioSimSpace#132 @richardjgowers
[joss] intro
re: openjournals/joss-reviews#1831

The intro to the paper is hard to read (for me at least), especially the first few sentences.  I'd recommend rewriting this to make it less general and nebulous.
