### 2019-11-10 #1831 @richardjgowers
Ok I've raised some minor issues but otherwise I think the paper is looking good
