### 2019-11-10 #1831 @lohedges
👍 Thanks, @richardjgowers. I agree with all of your suggestions and will make some edits tomorrow.
