### 2019-11-11 michellab/BioSimSpace#133 @amritagos
[JOSS Review] Minor paper changes
Part of https://github.com/openjournals/joss-reviews/issues/1831. I do think the paper is well written, however, certain parts might be reconsidered.

- Use of the word "fantastic" (abstract)

I was also surprised to not see any mention of or discussion of relative merits of other codes which work in a sort of similar manner, in terms of unifying different tools (though BioSimSpace has many novel features). In particular, mention of [ASE](https://wiki.fysik.dtu.dk/ase/) and [MD Traj](http://mdtraj.org/development/) or [Cubby](https://onlinelibrary.wiley.com/doi/10.1002/jcc.24312/abstract) might be useful in this context.

I would expect the above revisions before further considerations and look forward to your comments on the same.
