### 2019-11-11 michellab/BioSimSpace#133 @lohedges
Hi @amritagos, thanks for your comments and suggestions. I'll make some edits over the next few days and let you know when revisions are ready for approval.
