### 2019-11-11 fd7370c
Squashed code examples. [ref #130] [ref openjournals/joss-reviews#1831] ***NO_CI***
