### 2019-11-11 michellab/BioSimSpace#130 @lohedges
I've done some squashing and made the comments a little less verbose. The first example is still a little longer, since it introduces some functionality and concepts that aren't explained elsewhere. The second example is shorter since there is no need to re-comment functionality that has been explained previously.
