### 2019-11-11 698b720
Added authors involved in original software concept and design. [ref openjournals/joss-reviews#1831]  ***NO_CI***
