### 2019-11-11 8f45ac5
Fixed typos in docstrings. ***NO_CI***
