### 2019-11-11 5c76f5c
Remove confusing and redundant viewMolecules function.
