### 2019-11-11 784aa6c
Allow view to be created from a list of molecular input files.
