+++ python/BioSimSpace/Notebook/_view.py
@@ -38,6 +38,7 @@ from Sire import Mol as _SireMol
 from Sire import System as _SireSystem
 
 from BioSimSpace import _is_notebook, _isVerbose
+from BioSimSpace import IO as _IO
 from BioSimSpace.Process._process import Process as _Process
 from BioSimSpace._SireWrappers import System as _System
 
@@ -53,8 +54,10 @@ class View():
            handle : :class:`Process <BioSimSpace.Process>`, \
                     :class:`System <BioSimSpace._SireWrappers.System>` \
                     :class:`System <BioSimSpace._SireWrappers.Molecule>` \
-                    :class:`System <BioSimSpace._SireWrappers.Molecules>`
-               A handle to a process, system, molecule, or molecule container.
+                    :class:`System <BioSimSpace._SireWrappers.Molecules>` \
+                    str, [str]
+               A handle to a process, system, molecule, or molecule container,
+               or the path to molecular input file(s).
         """
 
         # Make sure we're running inside a Jupyter notebook.
@@ -64,8 +67,22 @@ class View():
 
         # Check the handle.
 
+        # Convert tuple to list.
+        if isinstance(handle, tuple):
+            handle = list(handle)
+
+        # Convert single string to list.
+        if isinstance(handle, str):
+            handle = [handle]
+
+        # List of strings (file paths).
+        if isinstance(handle, list) and all(isinstance(x, str) for x in handle):
+            system = _IO.readMolecules(handle)
+            self._handle = system._getSireObject()
+            self._is_process = False
+
         # BioSimSpace process.
-        if isinstance(handle, _Process):
+        elif isinstance(handle, _Process):
             self._handle = handle
             self._is_process = True
 
@@ -82,8 +99,9 @@ class View():
             except:
                 raise TypeError("The handle must be of type 'BioSimSpace.Process', "
                                 "'BioSimSpace._SireWrappers.System', "
-                                "'BioSimSpace._SireWrappers.Molecule', or "
-                                "'BioSimSpace._SireWrappers.Molecules'")
+                                "'BioSimSpace._SireWrappers.Molecule', "
+                                "'BioSimSpace._SireWrappers.Molecules', "
+                                "'str', or a list of 'str' types.")
 
         # Create a temporary workspace for the view object.
         self._tmp_dir = _tempfile.TemporaryDirectory()
