### 2019-11-11 michellab/BioSimSpace#130 @lohedges
I'll also squash the examples in the figures when I get a chance.
