### 2019-11-12 ce0ff73
Squashed examples in figures. [ref #130] [ref openjournals/joss-reviews#1831] ***NO_CI***
