### 2019-11-12 da2881d
Use a more recent GROMACS reference. ***NO_CI*** [closes #134]
