### 2019-11-12 2a15dae
Added citations to other interoperable software. [ref #131 #133 openjournals/joss-reviews#1831] ***NO_CI***
