+++ paper/paper.bib
@@ -64,3 +64,38 @@ year = {2005}
     howpublished = "\url{http://siremol.org}",
     note = "[Online; accessed 22-October-2019]"
 }
+
+@article{ref-INTERMOL,
+  title={Lessons learned from comparing molecular dynamics engines on the SAMPL5 dataset},
+  author={Shirts, Michael R and Klein, Christoph and Swails, Jason M and Yin, Jian and Gilson, Michael K and Mobley, David L and Case, David A and Zhong, Ellen D},
+  journal={Journal of computer-aided molecular design},
+  pages={1--15},
+  year={2016},
+  doi={doi:10.1007/s10822-016-9977-1}
+  publisher={Springer International Publishing}
+}
+
+@InProceedings{ref-MDANALYSIS,
+  author    = { {R}ichard {J}. {G}owers and {M}ax {L}inke and {J}onathan {B}arnoud and {T}yler {J}. {E}. {R}eddy and {M}anuel {N}. {M}elo and {S}ean {L}. {S}eyler and {J}an {D}omański and {D}avid {L}. {D}otson and {S}ébastien {B}uchoux and {I}an {M}. {K}enney and {O}liver {B}eckstein },
+  title     = { {M}{D}{A}nalysis: {A} {P}ython {P}ackage for the {R}apid {A}nalysis of {M}olecular {D}ynamics {S}imulations },
+  booktitle = { {P}roceedings of the 15th {P}ython in {S}cience {C}onference },
+  pages     = { 98 - 105 },
+  year      = { 2016 },
+  editor    = { {S}ebastian {B}enthall and {S}cott {R}ostrup },
+  doi       = { 10.25080/Majora-629e541a-00e }
+}
+
+@article{ref-CUBY,
+author = {Řezáč, Jan},
+title = {Cuby: An integrative framework for computational chemistry},
+journal = {Journal of Computational Chemistry},
+volume = {37},
+number = {13},
+pages = {1230-1237},
+keywords = {software framework, workflow automation, QM/MM, datasets, Ruby},
+doi = {10.1002/jcc.24312},
+url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/jcc.24312},
+eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1002/jcc.24312},
+abstract = {Cuby is a computational chemistry framework written in the Ruby programming language. It provides unified access to a wide range of computational methods by interfacing external software and it implements various protocols that operate on their results. Using structured input files, elementary calculations can be combined into complex workflows. For users, Cuby provides a unified and userfriendly way to automate their work, seamlessly integrating calculations carried out in different computational chemistry programs. For example, the QM/MM module allows combining methods across the interfaced programs and the builtin molecular dynamics engine makes it possible to run a simulation on the resulting potential. For programmers, it provides high-level, object-oriented environment that allows rapid development and testing of new methods and computational protocols. The Cuby framework is available for download at http://cuby4.molecular.cz. © 2016 Wiley Periodicals, Inc.},
+year = {2016}
+}
