### 2019-11-12 81170f6
Reworked "Summary" section. [ref #132 #133 openjournals/joss-reviews#1831] ***NO_CI***
