### 2019-11-12 b293e0c
Fixed broken InterMol citation (which was downloaded from the DOI). ***NO_CI***
