+++ paper/paper.bib
@@ -71,7 +71,7 @@ year = {2005}
   journal={Journal of computer-aided molecular design},
   pages={1--15},
   year={2016},
-  doi={doi:10.1007/s10822-016-9977-1}
+  doi={doi:10.1007/s10822-016-9977-1},
   publisher={Springer International Publishing}
 }
 
