### 2019-11-12 #1831 @lohedges
Thanks, @richardjgowers and @amritagos for your helpful comments. I've now addressed all of the issues and an updated PDF is available above. Conveniently, the changes have much improved the layout of the paper: code snippets are no longer split across multiple pages and figures flow nicely with the text.

I've cross-referenced commits with the issues raised and leave it up to the reviewers to close the respective issues when they are satisfied with the changes.

From discussion with my colleagues, we have also added several additional authors who were instrumental in the original concept and design of BioSimSpace. We have circulated the PDF to them and are awaiting comment.

Cheers.
