### 2019-11-12 #1831 @jmichel80
Excellent work @lohedges ! I will give it a read through later today and make minor edits.

--------------------------------------------------------------
Dr. Julien Michel,
Senior Lecturer
Room 263, School of Chemistry
University of Edinburgh
David Brewster road
Edinburgh, EH9 3FJ
United Kingdom
phone:   +44 (0)131 650 4797
http://www.julienmichel.net/
-------------------------------------------------------------
