+++ paper/paper.bib
@@ -99,3 +99,50 @@ eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1002/jcc.24312},
 abstract = {Cuby is a computational chemistry framework written in the Ruby programming language. It provides unified access to a wide range of computational methods by interfacing external software and it implements various protocols that operate on their results. Using structured input files, elementary calculations can be combined into complex workflows. For users, Cuby provides a unified and userfriendly way to automate their work, seamlessly integrating calculations carried out in different computational chemistry programs. For example, the QM/MM module allows combining methods across the interfaced programs and the builtin molecular dynamics engine makes it possible to run a simulation on the resulting potential. For programmers, it provides high-level, object-oriented environment that allows rapid development and testing of new methods and computational protocols. The Cuby framework is available for download at http://cuby4.molecular.cz. © 2016 Wiley Periodicals, Inc.},
 year = {2016}
 }
+
+@article{ref-CCPBioSimWIRES,
+author = {David J. Huggins,  Philip C. Biggin,  Marc A. Dämgen,  Jonathan W. Essex,  Sarah A. Harris,  Richard H. Henchman,  Syma Khalid,  Antonija Kuzmanic, Charles A. Laughton,  Julien Michel,  Adrian J. Mulholland,  Edina Rosta,  Mark S. P. Sansom,  Marc W. van der Kamp},
+title = {Biomolecular simulations: from dynamics and mechanisms to computational assays of biological activity},
+journal = {Wiley Interdisciplinary Reviews: Computational Molecular Science},
+volume = {9},
+number = {3},
+pages = {e1393},
+keywords = {enzyme membrane molecular dynamics multiscale protein QM/MM},
+doi = {10.1002/wcms.1393},
+url = {https://onlinelibrary.wiley.com/doi/full/10.1002/wcms.1393},
+eprint = {https://onlinelibrary.wiley.com/doi/full/10.1002/wcms.1393},
+abstract = {Biomolecular simulation is increasingly central to understanding and designing biological molecules and their interactions. Detailed, physics‐based simulation methods are demonstrating rapidly growing impact in areas as diverse as biocatalysis, drug delivery, biomaterials, biotechnology, and drug design. Simulations offer the potential of uniquely detailed, atomic‐level insight into mechanisms, dynamics, and processes, as well as increasingly accurate predictions of molecular properties. Simulations can now be used as computational assays of biological activity, for example, in predictions of drug resistance. Methodological and algorithmic developments, combined with advances in computational hardware, are transforming the scope and range of calculations. Different types of methods are required for different types of problem. Accurate methods and extensive simulations promise quantitative comparison with experiments across biochemistry. Atomistic simulations can now access experimentally relevant timescales for large systems, leading to a fertile interplay of experiment and theory and offering unprecedented opportunities for validating and developing models. Coarse‐grained methods allow studies on larger length‐ and timescales, and theoretical developments are bringing electronic structure calculations into new regimes. Multiscale methods are another key focus for development, combining different levels of theory to increase accuracy, aiming to connect chemical and molecular changes to macroscopic observables. In this review, we outline biomolecular simulation methods and highlight examples of its application to investigate questions in biology.},
+year = {2019}
+}
+
+@article{ref-reprodDDG,
[…]
