### 2019-11-12 eb5bc0b
Fixed citation formatting. ***NO_CI***
