### 2019-11-12 f982fc2
Remove duplicate use of "for instance" and grammar tweak. ***NO_CI***
