### 2019-11-13 b256af8
Added citation for MDTraj and ParmEd. ***NO_CI***
