+++ paper/paper.bib
@@ -145,4 +145,24 @@ abstract = {},
 year = {2015}
 }
 
+@article{ref-MDTRAJ,
+    title = {MDTraj: A Modern Open Library for the Analysis of
+    Molecular Dynamics Trajectories},
+    author = {McGibbon, Robert T. and Beauchamp, Kyle A. and Harrigan,
+    Matthew P. and Klein, Christoph and Swails, Jason M. and
+    Hern{\'a}ndez, Carlos X.  and Schwantes, Christian R. and Wang,
+    Lee-Ping and Lane, Thomas J. and Pande, Vijay S.},
+    journal = {Biophysical Journal},
+    volume = {109},
+    number = {8},
+    pages = {1528 -- 1532},
+    year = {2015},
+    doi = {10.1016/j.bpj.2015.08.015}
+}
 
+@misc{ref-PARMED,
+    author = "{Swails, Jason}",
+    title = {ParmEd: Cross-program parameter and topology file editor and molecular mechanical simulator engine},
+    year = "2010",
+    howpublished = "\url{https://github.com/ParmEd/ParmEd}"
+}
