### 2019-11-13 0cc98eb
Added ACPYPE citation. ***NO_CI***
