+++ paper/paper.bib
@@ -166,3 +166,30 @@ year = {2015}
     year = "2010",
     howpublished = "\url{https://github.com/ParmEd/ParmEd}"
 }
+
+@Article{ref-ACPYPE,
+author={Sousa da Silva, Alan W.
+and Vranken, Wim F.},
+title={ACPYPE - AnteChamber PYthon Parser interfacE},
+journal={BMC research notes},
+year={2012},
+month={Jul},
+day={23},
+publisher={BioMed Central},
+volume={5},
+pages={367-367},
+keywords={Computational Biology/*methods},
+keywords={Databases, Protein},
+keywords={Magnetic Resonance Spectroscopy},
+keywords={*Programming Languages},
+keywords={Small Molecule Libraries/analysis},
+keywords={*Software},
+keywords={*User-Computer Interface},
+abstract={BACKGROUND: ACPYPE (or AnteChamber PYthon Parser interfacE) is a wrapper script around the ANTECHAMBER software that simplifies the generation of small molecule topologies and parameters for a variety of molecular dynamics programmes like GROMACS, CHARMM and CNS. It is written in the Python programming language and was developed as a tool for interfacing with other Python based applications such as the CCPN software suite (for NMR data analysis) and ARIA (for structure calculations from NMR data). ACPYPE is open source code, under GNU GPL v3, and is available as a stand-alone application at http://www.ccpn.ac.uk/acpype and as a web portal application at http://webapps.ccpn.ac.uk/acpype. FINDINGS: We verified the topologies generated by ACPYPE in three ways: by comparing with default AMBER topologies for standard amino acids; by generating and verifying topologies for a large set of ligands from the PDB; and by recalculating the structures for 5 protein-ligand complexes from the PDB. CONCLUSIONS: ACPYPE is a tool that simplifies the automatic generation of topology and parameters in different formats for different molecular mechanics programmes, including calculation of partial charges, while being object oriented for integration with other applications.},
+note={22824207[pmid]},
+note={PMC3461484[pmcid]},
+note={1756-0500-5-367[PII]},
+issn={1756-0500},
+doi={10.1186/1756-0500-5-367},
+url={https://www.ncbi.nlm.nih.gov/pubmed/22824207}
+}
