### 2019-11-13 bf2ae6a
Updated references. ***NO_CI***
