+++ paper/paper.bib
@@ -9,7 +9,6 @@ keywords = {Amber, biomolecular simulation programs},
 doi = {10.1002/jcc.20290},
 url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/jcc.20290},
 eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1002/jcc.20290},
-abstract = {Abstract We describe the development, current features, and some directions for future development of the Amber package of computer programs. This package evolved from a program that was constructed in the late 1970s to do Assisted Model Building with Energy Refinement, and now contains a group of programs embodying a number of powerful tools of modern computational chemistry, focused on molecular dynamics and free energy calculations of proteins, nucleic acids, and carbohydrates. © 2005 Wiley Periodicals, Inc. J Comput Chem 26: 1668–1688, 2005},
 year = {2005}
 }
 
@@ -24,7 +23,6 @@ doi = "https://doi.org/10.1016/j.softx.2015.06.001",
 url = "http://www.sciencedirect.com/science/article/pii/S2352711015000059",
 author = "Mark James Abraham and Teemu Murtola and Roland Schulz and Szilárd Páll and Jeremy C. Smith and Berk Hess and Erik Lindahl",
 keywords = "Molecular dynamics, GPU, SIMD, Free energy",
-abstract = "GROMACS is one of the most widely used open-source and free software codes in chemistry, used primarily for dynamical simulations of biomolecules. It provides a rich set of calculation types, preparation and analysis tools. Several advanced techniques for free-energy calculations are supported. In version 5, it reaches new performance heights, through several new and enhanced parallelization algorithms. These work on every level; SIMD registers inside cores, multithreading, heterogeneous CPU–GPU acceleration, state-of-the-art 3D domain decomposition, and ensemble-level parallelization through built-in replica exchange and the separate Copernicus framework. The latest best-in-class compressed trajectory storage format is supported."
 }
 
 @article{ref-NAMD,
@@ -38,7 +36,6 @@ keywords = {biomolecular simulation, molecular dynamics, parallel computing},
 doi = {10.1002/jcc.20289},
 url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/jcc.20289},
 eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1002/jcc.20289},
[…]
