### 2019-11-13 719151d
Added "Advanced simulation methods" section. ***NO_CI***
