### 2019-11-13 8533eee
Remove early reference to "nodes" before they are define. ***NO_CI***
