### 2019-11-14 5be1641
Use hyperlinks since these are rendered in the published PDF. ***NO_CI***
