### 2019-11-14 0fbc29c
Added links for ipywidgets and argparse. ***NO_CI***
