### 2019-11-14 e4d7ff5
Removed line break. ***NO_CI***
