### 2019-11-14 2ad4b0d
Update paper.md
