### 2019-11-14 3225164
Formatting tweaks. ***NO_CI***
