+++ paper/paper.md
@@ -229,8 +229,8 @@ BioSimSpace has been developed with the intention of being easily extensible. Ad
 
 ## Advanced simulation methods
 
-As well as the basic molecular dynamics protocols described so far, BioSimSpace also supports several advanced biomolecular simulation techniques that can be deployed with modular pipelines of setup, run, and analysis nodes executed with best-practices protocols encoded in the library. Our intention is to make it easier to benchmark complex biomolecular simulation techniques by varying different setup tools, simulation engines or analysis techniques. 
-For instance Free-Energy Perturbation (FEP) [@ref-FEP] functionality in BioSimSpace can currently be used to compute drug binding affinities with the simulation engines SOMD or GROMACS, keeping all setup and analysis protocols identicals so that any variability in the results can be ascribed to differences in the simulation engines and protocols only. BioSimSpace also provides support for metadynamics [@ref-METADYNAMICS] simulations using [PLUMED](https://www.plumed.org) [@ref-PLUMED] and GROMACS. The application of BioSimSpace FEP and metadynamics workflows to proteins of pharmaceutical interest will be reported elsewhere in due course. 
+As well as the basic molecular dynamics protocols described so far, BioSimSpace also supports several advanced biomolecular simulation techniques that can be deployed with modular pipelines of setup, run, and analysis nodes executed with best-practices protocols encoded in the library. Our intention is to make it easier to benchmark complex biomolecular simulation techniques by varying different setup tools, simulation engines, or analysis techniques.
+For instance, Free-Energy Perturbation (FEP) [@ref-FEP] functionality in BioSimSpace can currently be used to compute drug binding affinities with the SOMD or GROMACS simulation engines. By keeping all setup and analysis protocols identical any variability in the results can be ascribed to differences in the simulation engines and protocols only. BioSimSpace also provides support for metadynamics [@ref-METADYNAMICS] simulations using [PLUMED](https://www.plumed.org) [@ref-PLUMED] and GROMACS. The application of BioSimSpace FEP and metadynamics workflows to proteins of pharmaceutical interest will be reported elsewhere in due course.
 
 ## Ease of use
 
