### 2019-11-14 a66b678
Restrict size of ring to speed up search.
