+++ paper/paper.md
@@ -51,7 +51,7 @@ Biomolecular simulation is a diverse and growing area of research, making import
 
 Several existing software packages attempt to address this problem: [InterMol](https://github.com/shirtsgroup/InterMol) [@ref-INTERMOL] and [ParmEd](https://github.com/ParmEd/ParmEd) [@ref-PARMED] can be used to read and write a wide variety of common molecular file formats; [ACPYPE](https://www.ccpn.ac.uk/v2-software/software/ACPYPE-folder) [@ref-ACPYPE] can generate small molecule topologies and parameters for a variety of molecular dynamics engines; [MDTraj](http://mdtraj.org) [@ref-MDTRAJ] and [MDAnalysis](https://www.mdanalysis.org) [@ref-MDANALYSIS] support reading, writing, and analysis of different molecular trajectory formats; the [Atomic Simulation Engine](https://wiki.fysik.dtu.dk/ase) (ASE) handles a wide variety of atomistic simulation tasks and provides interfaces to a range of external packages; and the [Cuby](http://cuby4.molecular.cz) [@ref-CUBY] framework allows access to a range of computational chemistry functionality from external packages, which can be combined into complex workflows through structured input files. Despite their utility, the above packages either have a restricted domain of application, e.g. trajectory files, or require different configuration options or scripts to interface with different external packages. It is not possible to write a _single_ script that is _independent_ of the underlying software packages installed on the host system.
 
-Within the Collaborative Computational Project for Biomolecular Simulation ([CCPBioSim](http://www.ccpbiosim.ac.uk)), we have attempted to solve this problem via the introduction of an interoperable framework, [BioSimSpace](https://biosimspace.org), that collects together the core functionality of many packages and exposes it through a simple API. By not choosing to reinvent the wheel, we can take advantage of all the exsting software within the community, and can easily plug into new software packages as they appear. Our software can convert between many common molecular file formats and automatically find packages available within the environment on which it is run. This allows the user to write portable workflow components that can be run with different input, on different environments, and in completely different ways, e.g. from the command-line, or within a [Jupyter](https://jupyter.org) notebook running on a cloud server. BioSimSpace builds on ideas explored previously by CCPBioSim during the development of the alchemical free energy calculations software [FESetup](https://github.com/CCPBioSim/fesetup) [@ref-FESetup] that provides consistent setup of input files for several simulation engines.
+Within the Collaborative Computational Project for Biomolecular Simulation ([CCPBioSim](http://www.ccpbiosim.ac.uk)), we have attempted to solve this problem via the introduction of an interoperable framework, [BioSimSpace](https://biosimspace.org), that collects together the core functionality of many packages and exposes it through a simple Python API. By not choosing to reinvent the wheel, we can take advantage of all the exsting software within the community, and can easily plug into new software packages as they appear. Our software can convert between many common molecular file formats and automatically find packages available within the environment on which it is run. This allows the user to write portable workflow components that can be run with different input, on different environments, and in completely different ways, e.g. from the command-line, or within a [Jupyter](https://jupyter.org) notebook running on a cloud server. BioSimSpace builds on ideas explored previously by CCPBioSim during the development of the alchemical free energy calculations software [FESetup](https://github.com/CCPBioSim/fesetup) [@ref-FESetup] that provides consistent setup of input files for several simulation engines.
 
 # Molecular dynamics
 
