### 2019-11-15 #1831 @lohedges
Apologies for all the re-renderings. I've just noticed that the [preview service](https://whedon.theoj.org) now properly supports custom branches (this wasn't the case when I first submitted) so I should have tested there instead.

I've now incorporated some minor changes suggested by my co-authors and have added page breaks to improve the paper formatting. For reference (in case anyone else searches for the same issue) the correct command to use is:

```latex
\pagebreak
```

rather than the HTML suggested previously:

```html
<div style="page-break-after: always;"></div>
```
(I assume this is because Pandoc is being used to generate the PDF.)

If changes are approved then I'd like to merge the `feature-paper` branch into our `devel` branch then create a new BioSimSpace release (2019.3.0) for the paper publication. (There have also been a few minor edits to `devel` since the review process started to fix typos in the documentation.) I assume that this means the version number, e.g. at the top of the review thread, would need to be updated.
