### 2019-11-15 3e53db8
Fixed citations. ***NO_CI***
