+++ paper/paper.bib
@@ -96,45 +96,45 @@ year = {2016}
 }
 
 @article{ref-CCPBioSimWIRES,
-author = {David J. Huggins,  Philip C. Biggin,  Marc A. Dämgen,  Jonathan W. Essex,  Sarah A. Harris,  Richard H. Henchman,  Syma Khalid,  Antonija Kuzmanic, Charles A. Laughton,  Julien Michel,  Adrian J. Mulholland,  Edina Rosta,  Mark S. P. Sansom,  Marc W. van der Kamp},
-title = {Biomolecular simulations: from dynamics and mechanisms to computational assays of biological activity},
+author = {Huggins, David J. and Biggin, Philip C. and Dämgen, Marc A. and Essex, Jonathan W. and Harris, Sarah A. and Henchman, Richard H. and Khalid, Syma and Kuzmanic, Antonija and Laughton, Charles A. and Michel, Julien and Mulholland, Adrian J. and Rosta, Edina and Sansom, Mark S. P. and van der Kamp, Marc W.},
+title = {Biomolecular simulations: From dynamics and mechanisms to computational assays of biological activity},
 journal = {Wiley Interdisciplinary Reviews: Computational Molecular Science},
 volume = {9},
 number = {3},
 pages = {e1393},
-keywords = {enzyme membrane molecular dynamics multiscale protein QM/MM},
+keywords = {enzyme, membrane, molecular dynamics, multiscale, protein, QM/MM},
 doi = {10.1002/wcms.1393},
-url = {https://onlinelibrary.wiley.com/doi/full/10.1002/wcms.1393},
-eprint = {https://onlinelibrary.wiley.com/doi/full/10.1002/wcms.1393},
+url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/wcms.1393},
+eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1002/wcms.1393},
 year = {2019}
 }
 
-@article{ref-reprodDDG,
-author = {Hannes H. Loeffler, Stefano Bosisio, Guilherme Duarte Ramos Matos, Donghyuk Suh, Benoit Roux, David L. Mobley, Julien Michel},
-title = {Reproducibility of Free Energy Calculations Across Different Molecular Simulation Software},
-journal = {Journal of Chemical Theory and Computation},
-volume = {14},
-number = {11},
-pages = {5567-5582},
-keywords = {},
-doi = {10.1021/acs.jctc.8b00544},
-url = {http://dx.doi.org/10.1021/acs.jctc.8b00544},
-eprint = {http://dx.doi.org/10.1021/acs.jctc.8b00544},
-year = {2018}
-}
-
 @article{ref-FESetup,
-author = {Hannes H. Loeffler, Julien Michel, Christopher Woods},
+author = {Loeffler, Hannes H. and Michel, Julien and Woods, Christopher J.},
 title = {FESetup: Automating Setup for Alchemical Free Energy Simulations},
 journal = {Journal of Chemical Information and Modeling},
 volume = {55},
 number = {12},
 pages = {2485-2490},
-keywords = {},
+year = {2015},
 doi = {10.1021/acs.jcim.5b00368},
-url = {http://dx.doi.org/10.1021/acs.jcim.5b00368},
-eprint = {http://dx.doi.org/10.1021/acs.jcim.5b00368},
-year = {2015}
+note = {PMID: 26544598},
+URL = {https://doi.org/10.1021/acs.jcim.5b00368},
+eprint = {https://doi.org/10.1021/acs.jcim.5b00368}
+}
+
+@article{ref-reprodDDG,
[…]
