### 2019-11-15 #1831 @lohedges
Apologies again. Some formatting issues in some of the citations that were only just spotted by a co-author. (The joy of hand-edited BibTex entries!)
