### 2019-11-15 michellab/BioSimSpace#133 @amritagos
The changes are very good, however on my final read I noticed something which might be clarified a little further.

![image](https://user-images.githubusercontent.com/32496815/68948681-516da480-07b0-11ea-9f37-163dbc5dea44.png)

What guesses are made in general when it comes to reading the file formats? Perhaps this section could be expanded a little bit or it might link to further documentation?
