### 2019-11-15 #1831 @lohedges
Just wondering why there appear to be a bunch of recent edits to the original post by @HaoZeke, who isn't an editor or reviewer of this paper?
