### 2019-11-15 michellab/BioSimSpace#133 @lohedges
I'll have a think how it might be best to clarify this. The guesses are very specific to individual parsers and we only use them when they are reliable. For example, we can reliably infer missing element records from the atom names in a PDB file, but don't attempt to account for missing atoms, or incorrect residue names.
