### 2019-11-15 #1831 @HaoZeke
Oh, that is odd. I haven't seen this before. I guess @amritagos used a lab PC, I might have been signed in when I visited. I will clear my login sessions. Shouldn't happen again.
