### 2019-11-15 michellab/BioSimSpace#133 @chryswoods
We don’t guess. Only information that is in the file is packed into the molecule, and thus only this information can be unpacked and written back out. This is part of what we mean by the parsers being symmetrical (same in, same out). The **only** inferred information is when reading a PDB when we follow the conventions of the field and assign chemical element information based on atom name. This is something I didn’t like adding, but is a necessary convention because of the way the field uses PDBs and overloads meaning into the atom name. Tools would simply break if we discarded this overloaded implied information.
