### 2019-11-16 michellab/BioSimSpace#133 @lohedges
When writing I was thinking about the PDB case and also certain situations when we need to make a best guess when information in an input file is poorly formatted. For example, if atom names (or ordering) aren't consistent between a GROMACS top and crd file. This isn't really inferring missing information, so I'll remove that from the text to avoid confusion.
