### 2019-11-16 6afe6a6
Remove suggestion that we might guess missing information. [ref #133] ***NO_CI***
