### 2019-11-18 michellab/BioSimSpace#133 @amritagos
Thanks. I am pleased with the changes and am happy to close this now.
