### 2019-11-22 #1831 @will-rowe
Great, thanks a lot @richardjgowers!

@lohedges - looks like we are ready to accept this. Before I start the pre-publication steps, can you carry out the branch changes you mentioned earlier? Once you have done this, can you then tag a new release and archive it? We'll need the DOI and version number
