### 2019-11-22 #1831 @lohedges
Thanks, everyone. I'll try to get those bits done this afternoon.

One quick grammatical note: Do you know whether JOSS prefer's capitalisation after the colon, i.e. as is done here? Looking at the list of [published papers](https://joss.theoj.org/papers/published) it looks like there is a mixture of styles. I can't see it mentioned anywhere.

BioSimSpace: An interoperable Python framework for biomolecular simulation
BioSimSpace: an interoperable Python framework for biomolecular simulation
