### 2019-11-22 #1831 @lohedges
I guess the example publication [here](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography) capitalises, so I'll stick with our title.
