### 2019-11-22 d6d7782
Update changelog for the 2019.3.0 release.
