### 2019-11-22 #1831 @lohedges
Hi @will-rowe, all sorted. I've merged into `devel`, created a [2019.3.0 release](https://github.com/michellab/BioSimSpace/releases/tag/2019.3.0), then merged across to our `master` (release) branch.  [This](http://doi.org/10.5281/zenodo.3551030) is the DOI for our 2019.3.0 Zenodo software archive.

Note that I've now deleted the redundant `feature-paper` branch, so the final publication will need to be compiled from `devel` (our default branch, so you can just use the [main GitHub repo link](https://github.com/michellab/BioSimSpace)), or `master`.

Cheers.
