### 2019-11-22 #1831 @will-rowe
Hey @openjournals/joss-eics - I think this one is good to go.
