### 2019-11-22 #1831 @kthyng
Congrats to @lohedges on your new paper!! Thanks for your time and expertise to editor @will-rowe and reviewers @amritagos and @richardjgowers. We couldn't have done this without you all!!
