### 2019-11-23 #1831 @lohedges
Thanks, everyone. This whole process has been a real breath of fresh air. Have a good weekend!
